+++ src/command/CommandMoveAnimationKey.hpp
@@ -15,7 +15,7 @@ class CommandMoveAnimationKey : public BaseCommand {
 public:
     // Constructor: Move key by cellanimIndex, animationIndex and keyIndex.
     CommandMoveAnimationKey(
-        uint16_t cellanimIndex, uint16_t animationIndex, uint32_t keyIndex,
+        unsigned cellanimIndex, unsigned animationIndex, unsigned keyIndex,
         bool moveDown, bool preserveHold
     ) :
         cellanimIndex(cellanimIndex), animationIndex(animationIndex), keyIndex(keyIndex),
@@ -27,8 +27,8 @@ public:
 
         RvlCellAnim::Animation& animation = this->getAnimation();
 
-        // An integer is used since nSwap can be negative.
-        int32_t nSwap = this->keyIndex + (moveDown ? -1 : 1);
+        // An signed int is used since nSwap can be negative.
+        int nSwap = this->keyIndex + (moveDown ? -1 : 1);
         if (nSwap >= 0 && nSwap < animation.keys.size()) {
             std::swap(
                 animation.keys.at(this->keyIndex),
@@ -51,7 +51,7 @@ public:
 
         RvlCellAnim::Animation& animation = this->getAnimation();
 
-        int32_t nSwap = this->keyIndex + (moveDown ? -1 : 1);
+        int nSwap = this->keyIndex + (moveDown ? -1 : 1);
         if (nSwap >= 0 && nSwap < animation.keys.size()) {
             std::swap(
                 animation.keys.at(this->keyIndex),
@@ -76,9 +76,9 @@ private:
 
     bool preserveHold;
 
-    uint16_t cellanimIndex;
-    uint16_t animationIndex;
-    uint32_t keyIndex;
+    unsigned cellanimIndex;
+    unsigned animationIndex;
+    unsigned keyIndex;
 
     RvlCellAnim::Animation& getAnimation() {
         return
